improve: HQM-153: Remove useWalletsList and useWalletsVisible
