+++ src/hooks/index.ts
@@ -1,8 +1,6 @@
 export * from './use-typed-navigation';
 export * from './use-theme';
 export * from './use-typed-route';
-export * from './use-wallets-list';
-export * from './use-wallets-visible';
 export * from './use-cosmos';
 export * from './use-sum-amount';
 export * from './use-keyboard-dismiss-in-background';
