+++ src/screens/home-staking.tsx
@@ -9,11 +9,12 @@ import {onTrackEvent} from '@app/event-actions/on-track-event';
 import {Events} from '@app/events';
 import {abortProviderInstanceForWallet} from '@app/helpers/provider-instance';
 import {sumReduce} from '@app/helpers/staking';
-import {useTypedNavigation, useWalletsVisible} from '@app/hooks';
+import {useTypedNavigation} from '@app/hooks';
 import {
   StakingMetadata,
   StakingMetadataType,
 } from '@app/models/staking-metadata';
+import {Wallet} from '@app/models/wallet';
 import {Balance} from '@app/services/balance';
 import {AdjustEvents} from '@app/types';
 
@@ -25,7 +26,7 @@ const initData = {
 };
 
 export const HomeStakingScreen = observer(() => {
-  const visible = useWalletsVisible();
+  const visible = Wallet.getAllVisible();
 
   const [data, setData] = useState({
     ...initData,
