+++ src/screens/locked-tokens.tsx
@@ -4,12 +4,13 @@ import {observer} from 'mobx-react';
 
 import {LockedTokens} from '@app/components/locked-tokens';
 import {Feature, isFeatureEnabled} from '@app/helpers/is-feature-enabled';
-import {useTypedNavigation, useWalletsVisible} from '@app/hooks';
+import {useTypedNavigation} from '@app/hooks';
 import {useWalletsBalance} from '@app/hooks/use-wallets-balance';
+import {Wallet} from '@app/models/wallet';
 import {calculateBalances} from '@app/utils';
 
 export const LockedTokensWrapper = observer(() => {
-  const visible = useWalletsVisible();
+  const visible = Wallet.getAllVisible();
   const balances = useWalletsBalance(visible);
   const calculatedBalance = useMemo(
     () => calculateBalances(balances, visible),
