+++ src/screens/total-value-info.tsx
@@ -5,7 +5,7 @@ import {observer} from 'mobx-react';
 import {TotalValueInfo} from '@app/components/total-value-info';
 import {Loading} from '@app/components/ui';
 import {showModal} from '@app/helpers';
-import {useTypedNavigation, useWalletsVisible} from '@app/hooks';
+import {useTypedNavigation} from '@app/hooks';
 import {useEffectAsync} from '@app/hooks/use-effect-async';
 import {useTransactionList} from '@app/hooks/use-transaction-list';
 import {useWalletsBalance} from '@app/hooks/use-wallets-balance';
@@ -20,7 +20,7 @@ import {calculateBalances} from '@app/utils';
 
 export const TotalValueInfoScreen = observer(() => {
   const navigation = useTypedNavigation();
-  const wallets = useWalletsVisible();
+  const wallets = Wallet.getAllVisible();
   const adressList = useMemo(() => Wallet.addressList(), []);
   const transactionsList = useTransactionList(adressList);
   const balances = useWalletsBalance(wallets);
