+++ src/screens/transaction-account.tsx
@@ -3,8 +3,9 @@ import React, {useCallback} from 'react';
 import {observer} from 'mobx-react';
 
 import {TransactionAccount} from '@app/components/transaction-account';
-import {useTypedNavigation, useTypedRoute, useWalletsList} from '@app/hooks';
+import {useTypedNavigation, useTypedRoute} from '@app/hooks';
 import {useAndroidBackHandler} from '@app/hooks/use-android-back-handler';
+import {Wallet} from '@app/models/wallet';
 
 export const TransactionAccountScreen = observer(() => {
   const navigation = useTypedNavigation();
@@ -13,7 +14,7 @@ export const TransactionAccountScreen = observer(() => {
     return true;
   }, [navigation]);
   const route = useTypedRoute<'transactionAccount'>();
-  const wallets = useWalletsList();
+  const wallets = Wallet.getAll();
   const onPressRow = useCallback(
     (address: string) => {
       navigation.navigate('transactionAddress', {
