+++ src/screens/transaction-sum-address.tsx
@@ -4,9 +4,10 @@ import {observer} from 'mobx-react';
 
 import {TransactionAddress} from '@app/components/transaction-address';
 import {app} from '@app/contexts';
-import {useTypedNavigation, useTypedRoute, useWalletsVisible} from '@app/hooks';
+import {useTypedNavigation, useTypedRoute} from '@app/hooks';
 import {useAndroidBackHandler} from '@app/hooks/use-android-back-handler';
 import {Contact} from '@app/models/contact';
+import {Wallet} from '@app/models/wallet';
 
 export const TransactionSumAddressScreen = observer(() => {
   const navigation = useTypedNavigation();
@@ -15,7 +16,7 @@ export const TransactionSumAddressScreen = observer(() => {
     return true;
   }, [navigation]);
   const route = useTypedRoute<'transactionSumAddress'>();
-  const wallets = useWalletsVisible();
+  const wallets = Wallet.getAllVisible();
   const contacts = useRef(Contact.getAll()).current;
 
   const [address, setAddress] = useState(route.params?.to || '');
